+++ CHANGES.rst
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+Added
+*****
+
+- OIDC `prompt=create` support. :issue:`185` :pr:`164`
+
 Fixed
 *****
 
@@ -13,6 +18,7 @@ Fixed
 
 Added
 *****
+
 - ``THEME`` can be a relative path
 
 [0.0.39] - 2023-12-15